Switch to a `string.replace` approach
This approach is vastly simpler and faster, since it leverages the regular expression engine to to most of the work.

To handle comments inside of strings, simply search for strings as well, but then leave the strings as-is in the replacer function.

The benchmarks do run quite a bit faster:

- strip JSON comments is 80% faster
- strip JSON comments without whitespace is 180% faster
- strip Big JSON comments is 50% faster